Refactor SlackController to remove error handling
The previous implementation of the SlackUrlVerification class raised
errors if initialize received bad input.  This caused the
SlackController to handle those errors with a begin and rescue.  This
could be a problem if any unexpected errors are raised.

This commit refactors SlackUrlVerification to handle bad data when
execute is called and instead of raising errors, it returns a bad
request.

The SlackController was then refactored to remove the error handling and
instead handle a bad response.

Several private methods were also extracted to capture specific logic
and dry out the two classes.